# docs/04-api/openapi.yaml
openapi: 3.1.0
info:
  title: DevHub API
  version: 1.0.0
  description: |
    DevHub API — project management for programming projects.
    Auth via httpOnly cookie (devhub_session). All endpoints JSON.
    The MCP endpoint (/mcp) is a Model Context Protocol streamable-HTTP server
    and is documented separately in docs/03-engineering/mcp-integration.md.
  contact:
    name: DevHub Project Owner
  license:
    name: TBD
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://devhub.example.com/api
    description: Production (TBD)
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created; session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive session cookie
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in; session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out and clear session cookie
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Cookie cleared
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      operationId: me
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: devhub_session
      description: JWT session cookie set by register/login
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL]
            message:
              type: string
            details:
              type: array
              items:
                type: object
              description: Field-level errors (zod)
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
      additionalProperties: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      additionalProperties: false
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
...